Allow `Dummy` (#109)
This PR allows a type to be `Dummy` and still be accepted by the type-checker. There are way too much fine-grained types in the archicad file, that we mostly do not need to be implemented (coloring of certain elements, site geometry, ...).

I assume that, when we import a file we only want to have the hierarchy (`project` -> `building` -> `storey` -> `space`) and the geometry (walls, slabs, roofs, windows, ...) and not the programs properties (`IFCPROPERTYSET`, `IFCPROPERTYSINGLEVALUE`, ...).

Reviewed-on: https://git.mtbld.de/MetabuildStudio/IFC4/pulls/109
Reviewed-by: robert.walter <robert.walter@metabuild.io>
Co-authored-by: Michael Hübner <michael.huebner@metabuild.de>
Co-committed-by: Michael Hübner <michael.huebner@metabuild.de>