Archicad success (#114)
I got the test to succeed. I reverted the changes to `Dummy` this morning after a discussion with Robert about removing the fail for tests. Now I bring it back in to the discussion again. If we remove the `fail` here the test succeeds in the sense that the file can be imported, while kind of ignoring the types that we don't have.

The next step of mine would be to create an easy to use loader, that one can parse the hierarchy and get our currently supported types out of the the file.

Reviewed-on: https://git.mtbld.de/MetabuildStudio/IFC4/pulls/114
Reviewed-by: robert.walter <robert.walter@metabuild.io>
Co-authored-by: Michael Hübner <michael.huebner@metabuild.de>
Co-committed-by: Michael Hübner <michael.huebner@metabuild.de>